+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.3
+
+- Added PaylikeException as the general paylike API response error
+- Fixed an issue where request function did not provide enough information about API errors
+
 ## 1.0.2
 
 - Fixed an issue with adding JSON body